feat: Allow setting path where keys are stored + only store private key
Currently, it's possible to configure paths used by pocket-id to store the SQLite DB, uploads, and the GeoDB data. However, keys were stored in the hardcoded path `data/keys`.

This PR introduces the env var `KEYS_PATH` which allows overriding where keys are stored. Its default value is `data/keys`.

Additionally, as I was implementing this, I realized that pocket-id was writing to disk both the private and public keys, and loading them independently. This was both unnecessary, since the public key can always be derived from the private one, and potentially a bit dangerous, as users could modify the public key on disk and pocket-id was not checking to make sure the two keys were related to each other.

PS: Also included a performance optimization where the ID of the key is pre-computed once and saved in the object, rather than re-computed every time, since the computation is expensive and the value is static