refactor: do not include test controller in production builds
Currently, the test controller is built into Pocket ID binaries and is enabled every time `APP_ENV` is any value but `production`.

The test controller is very powerful, as it includes a RESTful endpoint that allows nuking the entire database. While it is required for E2E tests, its existence outside of that can be considered a security risk.

This PR uses Go build tags to make sure that the test controller is only built when the application is compiled with `-tag "e2etest"`, which is not included in production builds