feat: auto-focus on the login buttons
On the /login and /authorize views, use auto-focus to make the "Login" and "Sign In" buttons (respectively) automatically focused

This helps when doing keyboard navigation. For example, imagine refreshing the page with CMD/CTRL+R, being sent to Pocket ID to re-authorize an app after a session expired. Your fingers were on the keyboard, but now you need to use your mouse to select "Sign In" or "Login" (or press tab multiple times and pay attention to where the focus is), and then go back to the fingerprint sensor to make the passkey work.

By setting auto-focus, you can just hit enter and go straight to the passkey.