feat: Encrypt private keys saved on disk and in DB
Fixes #580

Implements the design described here: https://github.com/pocket-id/pocket-id/issues/580#issuecomment-2918342289

- Adds support for storing private keys in the database
- Adds support for encrypting the private keys stored on disk (optional) or in the database (required)
- Adds support for running Pocket ID with in-memory only (ephemeral) keys

Adds these new config options:

- `KEYS_STORAGE`, which can be `file`, `database`, or `memory`. The default is `file` if no key encryption key is passed, or `database` if there's a key encryption key
- `KEYS_ENCRYPTION_KEY` allows passing a key to use to encrypt keys stored on disk or DB
- `KEYS_ENCRYPTION_KEY_FILE` is akin to `KEYS_ENCRYPTION_KEY` but contains the path to a file with the key to load

When keys are encrypted, they are tied to a specific instance of Pocket ID (identified by the InstanceID internal config setting).

Also note that keys are encrypted with AES-CBC-HMAC ("Encrypt-then-MAC", using a 256-bit key for encryption and SHA-512 for HMAC; additionally, uses PKCS#7 for padding). This is an authenticated cipher (AEAD) but that doesn't have the issues with nonce re-use like AES-GCM. Support for this isn't available out-of-the-box in the Go's standard library, so this uses code copied from the Dapr project (licensed under Apache2)