fix: Auth fails when client IP is empty on Postgres
Fixes #686

Auth fails with Postgres when the client IP is empty, for example when proxies connect to Pocket ID via a Unix Domain Socket (UDS).

This was due to the fact that in Postgres we store IPs with the INET column type, which doesn't accept empty values.

The fix involves:

- Converting the ip_address column to be nullable in Postgres
   - In SQLite, we were using a non-nullable TEXT. Since altering columns in SQLite is not possible, it has not changed there.
- Make the IpAddress type nullable in the model, and use a custom type that serializes to an empty string on SQLite and a NULL value in Postgres

Additionally, improved error handling throughout the process.