fix: ensure user inputs are normalized
User inputs such as names can contain UTF-8 sequences, and they may not be in a normalized form, so they may not show up as identical when they should be. This PR makes sure that all strings that are not identifiers (like UUIDs) are normalized before being stored in the database.