feat: support reading secret env vars from _FILE
Fixes #685

Env vars that contain secret values now can be read from file. To do that, pass the env var `*_FILE` containing the path to a file on disk. This works with Docker (and K8s) secrets too, since they are mounted as files.

These env vars are currently supported:

- EnvConfig:
   - `DB_CONNECTION_STRING_FILE`
   - `MAXMIND_LICENSE_KEY_FILE`
   - Additionally, `ENCRYPTION_KEY_FILE` was already available
- AppConfig (when UI configuration is disabled): all values that have the `sensitive` tag in the struct, which currently includes:
   - `SMTP_PASSWORD_FILE`
   - `LDAP_BIND_PASSWORD_FILE`