chore: minify background image
The current background image is a very large JPEG file that is ove 3MB in size. That is a lot of data, especially for users that are not in the LAN.

This PR converts the background image to a smaller one (1080x720, plenty for how it's displayed even on HiDpi screens) and uses the WebP format (supported by all browsers that support WebAuthn). This reduces the size of the background image to just 105KB.

Because we did not have code in place to replace existing images only if they were the default, I added some support code that checks if file on disk represent legacy default images and deletes them - but only if they were the default ones (hashes are hardcode in the application)